indexer: Recreate FTP connection on unexpected ftplib error
On long connections or with buggy FTP servers, the connection can lose
synchronization and the indexer can be unable to recover from it.
Therefore, on unexpected errors from the server, the connection is
recreated to allow indexation of the remaining folders.

FTP_Indexer does not inherit from FTP_Retry anymore. The FTP connection
is now referenced in an attribute instead.